extract solana-config-program-api and re-export it in solana-config-program
