+++ Cargo.toml
@@ -325,6 +325,7 @@ solana-cli-output = { path = "cli-output", version = "=2.0.0" }
 solana-client = { path = "client", version = "=2.0.0" }
 solana-compute-budget-program = { path = "programs/compute-budget", version = "=2.0.0" }
 solana-config-program = { path = "programs/config", version = "=2.0.0" }
+solana-config-program-api = { path = "programs/config-api", version = "=0.0.1" }
 solana-connection-cache = { path = "connection-cache", version = "=2.0.0", default-features = false }
 solana-core = { path = "core", version = "=2.0.0" }
 solana-cost-model = { path = "cost-model", version = "=2.0.0" }
