+++ programs/config/Cargo.toml
@@ -14,6 +14,7 @@ bincode = { workspace = true }
 chrono = { workspace = true, features = ["default", "serde"] }
 serde = { workspace = true }
 serde_derive = { workspace = true }
+solana-config-program-api = { workspace = true }
 solana-program-runtime = { workspace = true }
 solana-sdk = { workspace = true }
